Update 0.5.1
- Unnecessary logging finally removed
- Moving DataOption & DataOptionList objects
