+++ build.py
@@ -31,6 +31,7 @@ DATA = {
     "seaplayer/objects/MusicList.py": "seaplayer/objects/",
     "seaplayer/objects/Notification.py": "seaplayer/objects/",
     "seaplayer/objects/ProgressBar.py": "seaplayer/objects/",
+    "seaplayer/objects/DataOptions.py": "seaplayer/objects/",
     # * 1.2) Types
     "seaplayer/types": "seaplayer/types/",
     "seaplayer/types/__init__.py": "seaplayer/types/",
