+++ seaplayer/objects/__init__.py
@@ -2,6 +2,7 @@ from .Log import LogMenu
 from .Input import InputField
 from .Notification import Nofy, CallNofy
 from .ProgressBar import IndeterminateProgress
+from .DataOptions import DataOption, DataOptionList
 from .Image import AsyncImageLabel, StandartImageLabel
 from .MusicList import MusicListView, MusicListViewItem
 from .Configurate import ConfigurateListItem, ConfigurateList
